chore(codegen): #1098 split expr.rs — extract nanbox_inline, strings, v8_interop submodules (behavior-preserving)
