+++ crates/perry-codegen/src/expr/mod.rs
@@ -21,9 +21,7 @@ use crate::lower_string_method::{
     flatten_string_add_chain, lower_string_coerce_concat, lower_string_concat,
     lower_string_concat_chain, lower_string_self_append,
 };
-use crate::nanbox::{
-    double_literal, BIGINT_TAG_I64, POINTER_MASK_I64, POINTER_TAG_I64, STRING_TAG_I64,
-};
+use crate::nanbox::{double_literal, POINTER_MASK_I64};
 use crate::strings::StringPool;
 use crate::type_analysis::{
     compute_auto_captures, is_array_expr, is_bigint_expr, is_bool_expr, is_map_expr,
@@ -41,8 +39,11 @@ mod channel;
 mod helpers;
 mod i32_fast_path;
 mod index;
+mod nanbox_inline;
 mod object_literal;
+mod strings;
 mod url_helpers;
+mod v8_interop;
 mod write_barrier;
 
 #[allow(unused_imports)] // ChannelReduction kept reachable for surface stability
@@ -64,374 +65,17 @@ pub(crate) use helpers::{
     unbox_to_i64,
 };
 pub(crate) use index::lower_index_set_fast;
+pub(crate) use nanbox_inline::{
+    i32_bool_to_nanbox, nanbox_bigint_inline, nanbox_pointer_inline, nanbox_pointer_inline_pub,
+    nanbox_string_inline,
+};
 pub(crate) use object_literal::lower_object_literal;
+pub(crate) use strings::emit_string_literal_global;
 pub(crate) use url_helpers::lower_url_string_getter;
+pub(crate) use v8_interop::{
+    emit_v8_export_call, emit_v8_member_method_call, import_origin_suffix, try_static_class_name,
+};
 
-/// Inline NaN-box of a raw heap pointer with `POINTER_TAG`.
-pub(crate) fn nanbox_pointer_inline(blk: &mut LlBlock, ptr_i64: &str) -> String {
-    let tagged = blk.or(I64, ptr_i64, POINTER_TAG_I64);
-    blk.bitcast_i64_to_double(&tagged)
-}
-
-/// Inline NaN-box of a raw `BigIntHeader*` with `BIGINT_TAG`. Required
-/// for `typeof x === "bigint"` (which reads the tag byte), and for the
-/// runtime's dynamic-dispatch helpers (`js_dynamic_add` etc.) to
-/// recognize the value as a bigint at their check sites. Without this,
-/// literals like `5n` get tagged as `POINTER_TAG` and `typeof` reports
-/// `"object"` / arithmetic falls back to float and returns `NaN`.
-pub(crate) fn nanbox_bigint_inline(blk: &mut LlBlock, ptr_i64: &str) -> String {
-    let tagged = blk.or(I64, ptr_i64, BIGINT_TAG_I64);
-    blk.bitcast_i64_to_double(&tagged)
-}
-
-/// Issue #678: resolve the actual symbol-suffix for an imported name.
-///
-/// Re-export renames like `export { default as render } from './render.js'`
-/// mean the consumer sees `render` while the origin module emits
-/// `perry_fn_<origin>__default`. The `import_function_origin_names` map
-/// records this rename so every `perry_fn_<src>__<suffix>` construction
-/// site can pick the right suffix instead of the consumer-visible name.
-///
-/// Returns the override when present, falling back to `name` (the common
-/// case where origin name == consumer name). Used by every codegen
-/// path that builds a `perry_fn_<src>__<name>` extern symbol.
-pub(crate) fn import_origin_suffix<'a>(
-    origin_names: &'a std::collections::HashMap<String, String>,
-    name: &'a str,
-) -> &'a str {
-    origin_names.get(name).map(String::as_str).unwrap_or(name)
-}
-
-/// Issue #841: materialize a NUL-terminated rodata constant carrying
-/// `text`'s UTF-8 bytes and return the LLVM IR pointer expression that
-/// names it. Used by the named-import + namespace-import value-form
-/// lowerings to hand a stable `(ptr, len)` pair to the runtime helpers
-/// `js_node_submodule_export_as_function` /
-/// `js_node_submodule_namespace`. The label uses a per-invocation
-/// counter so multiple call sites in the same function don't collide.
-pub(crate) fn emit_string_literal_global(ctx: &mut FnCtx<'_>, text: &str) -> String {
-    let idx = ctx.typed_parse_counter;
-    ctx.typed_parse_counter += 1;
-    let global_name = format!("perry_node_submod_str_{}", idx);
-    let bytes = text.as_bytes();
-    let mut lit = String::with_capacity(bytes.len() + 4);
-    lit.push('c');
-    lit.push('"');
-    for &b in bytes {
-        if (32..127).contains(&b) && b != b'"' && b != b'\\' {
-            lit.push(b as char);
-        } else {
-            lit.push('\\');
-            lit.push_str(&format!("{:02X}", b));
-        }
-    }
-    lit.push_str("\\00\"");
-    ctx.typed_parse_rodata.push(format!(
-        "@{} = private unnamed_addr constant [{} x i8] {}",
-        global_name,
-        bytes.len() + 1,
-        lit
-    ));
-    format!("@{}", global_name)
-}
-
-/// Issue #678 followup: emit a `js_call_v8_export` bridge call for a name
-/// that resolves to a V8-fallback (interpreted) module.
-///
-/// Materializes per-call-site rodata constants for the module specifier
-/// and the export name (linker merges duplicates across translation units),
-/// stack-allocates an f64 args array, and emits the runtime call. Returns
-/// the SSA double register holding the NaN-boxed result.
-///
-/// Caller has already validated that `name` is in
-/// `ctx.import_function_v8_specifiers`; this helper just emits the lowering.
-pub(crate) fn emit_v8_export_call(
-    ctx: &mut FnCtx<'_>,
-    specifier: &str,
-    export_name: &str,
-    lowered_args: &[String],
-) -> String {
-    let idx = ctx.typed_parse_counter;
-    ctx.typed_parse_counter += 1;
-    let spec_global = format!("perry_v8_spec_{}", idx);
-    let name_global = format!("perry_v8_name_{}", idx);
-    let escape = |s: &str| -> String {
-        let bytes = s.as_bytes();
-        let mut lit = String::with_capacity(bytes.len() + 4);
-        lit.push('c');
-        lit.push('"');
-        for &b in bytes {
-            if (32..127).contains(&b) && b != b'"' && b != b'\\' {
-                lit.push(b as char);
-            } else {
-                lit.push('\\');
-                lit.push_str(&format!("{:02X}", b));
-            }
-        }
-        lit.push_str("\\00\"");
-        lit
-    };
-    let spec_bytes = specifier.as_bytes().len();
-    let name_bytes = export_name.as_bytes().len();
-    ctx.typed_parse_rodata.push(format!(
-        "@{} = private unnamed_addr constant [{} x i8] {}",
-        spec_global,
-        spec_bytes + 1,
-        escape(specifier)
-    ));
-    ctx.typed_parse_rodata.push(format!(
-        "@{} = private unnamed_addr constant [{} x i8] {}",
-        name_global,
-        name_bytes + 1,
-        escape(export_name)
-    ));
-
-    let argc = lowered_args.len();
-    let alloca_count = if argc == 0 { 1 } else { argc };
-    let blk = ctx.block();
-    let argc_lit = format!("{}", argc);
-    let spec_ptr = format!("@{}", spec_global);
-    let name_ptr = format!("@{}", name_global);
-    let spec_len_lit = format!("{}", spec_bytes);
-    let name_len_lit = format!("{}", name_bytes);
-
-    // Stack-allocate the args buffer (zero-len → still need a pointer; an
-    // `alloca [1 x double]` is well-formed in LLVM and never dereferenced
-    // because argc=0 in that branch of the runtime).
-    let args_slot = blk.fresh_reg();
-    blk.emit_raw(format!(
-        "{} = alloca [{} x double], align 8",
-        args_slot, alloca_count
-    ));
-    for (i, v) in lowered_args.iter().enumerate() {
-        let slot = blk.fresh_reg();
-        blk.emit_raw(format!(
-            "{} = getelementptr inbounds [{} x double], ptr {}, i64 0, i64 {}",
-            slot, alloca_count, args_slot, i
-        ));
-        blk.emit_raw(format!("store double {}, ptr {}, align 8", v, slot));
-    }
-
-    ctx.pending_declares.push((
-        "js_call_v8_export".to_string(),
-        DOUBLE,
-        vec![PTR, I64, PTR, I64, PTR, I64],
-    ));
-    let blk = ctx.block();
-    blk.call(
-        DOUBLE,
-        "js_call_v8_export",
-        &[
-            (PTR, &spec_ptr),
-            (I64, &spec_len_lit),
-            (PTR, &name_ptr),
-            (I64, &name_len_lit),
-            (PTR, &args_slot),
-            (I64, &argc_lit),
-        ],
-    )
-}
-
-/// Issue #818 (Effect.succeed pattern): emit a
-/// `js_call_v8_member_method(spec, member, method, args)` bridge call for a
-/// named V8 import used as a static-method receiver — `Effect.succeed(42)`
-/// where `Effect` is `import { Effect } from 'effect'`. The V8 module's
-/// top-level export `Effect` is itself a namespace-shaped object whose
-/// `.succeed` is the actual function; the existing `emit_v8_export_call`
-/// would mistakenly try to invoke `effect.succeed(...)` at the module root.
-pub(crate) fn emit_v8_member_method_call(
-    ctx: &mut FnCtx<'_>,
-    specifier: &str,
-    member: &str,
-    method: &str,
-    lowered_args: &[String],
-) -> String {
-    let idx = ctx.typed_parse_counter;
-    ctx.typed_parse_counter += 1;
-    let spec_global = format!("perry_v8_mspec_{}", idx);
-    let member_global = format!("perry_v8_mmember_{}", idx);
-    let method_global = format!("perry_v8_mmethod_{}", idx);
-    let escape = |s: &str| -> String {
-        let bytes = s.as_bytes();
-        let mut lit = String::with_capacity(bytes.len() + 4);
-        lit.push('c');
-        lit.push('"');
-        for &b in bytes {
-            if (32..127).contains(&b) && b != b'"' && b != b'\\' {
-                lit.push(b as char);
-            } else {
-                lit.push('\\');
-                lit.push_str(&format!("{:02X}", b));
-            }
-        }
-        lit.push_str("\\00\"");
-        lit
-    };
-    let spec_bytes = specifier.as_bytes().len();
-    let member_bytes = member.as_bytes().len();
-    let method_bytes = method.as_bytes().len();
-    ctx.typed_parse_rodata.push(format!(
-        "@{} = private unnamed_addr constant [{} x i8] {}",
-        spec_global,
-        spec_bytes + 1,
-        escape(specifier)
-    ));
-    ctx.typed_parse_rodata.push(format!(
-        "@{} = private unnamed_addr constant [{} x i8] {}",
-        member_global,
-        member_bytes + 1,
-        escape(member)
-    ));
-    ctx.typed_parse_rodata.push(format!(
-        "@{} = private unnamed_addr constant [{} x i8] {}",
-        method_global,
-        method_bytes + 1,
-        escape(method)
-    ));
-
-    let argc = lowered_args.len();
-    let alloca_count = if argc == 0 { 1 } else { argc };
-    let blk = ctx.block();
-    let argc_lit = format!("{}", argc);
-    let spec_ptr = format!("@{}", spec_global);
-    let member_ptr = format!("@{}", member_global);
-    let method_ptr = format!("@{}", method_global);
-    let spec_len_lit = format!("{}", spec_bytes);
-    let member_len_lit = format!("{}", member_bytes);
-    let method_len_lit = format!("{}", method_bytes);
-
-    let args_slot = blk.fresh_reg();
-    blk.emit_raw(format!(
-        "{} = alloca [{} x double], align 8",
-        args_slot, alloca_count
-    ));
-    for (i, v) in lowered_args.iter().enumerate() {
-        let slot = blk.fresh_reg();
-        blk.emit_raw(format!(
-            "{} = getelementptr inbounds [{} x double], ptr {}, i64 0, i64 {}",
-            slot, alloca_count, args_slot, i
-        ));
-        blk.emit_raw(format!("store double {}, ptr {}, align 8", v, slot));
-    }
-
-    ctx.pending_declares.push((
-        "js_call_v8_member_method".to_string(),
-        DOUBLE,
-        vec![PTR, I64, PTR, I64, PTR, I64, PTR, I64],
-    ));
-    let blk = ctx.block();
-    blk.call(
-        DOUBLE,
-        "js_call_v8_member_method",
-        &[
-            (PTR, &spec_ptr),
-            (I64, &spec_len_lit),
-            (PTR, &member_ptr),
-            (I64, &member_len_lit),
-            (PTR, &method_ptr),
-            (I64, &method_len_lit),
-            (PTR, &args_slot),
-            (I64, &argc_lit),
-        ],
-    )
-}
-
-/// If `callee` is a `new`-target whose class name is statically
-/// known, return that name. Used by the `Expr::NewDynamic` lowering
-/// to reroute statically-resolvable shapes to the regular `lower_new`
-/// path. Returns `None` for any callee that needs runtime dispatch
-/// (locals, conditionals with non-classy arms, computed expressions).
-///
-/// Recognized shapes:
-///   - `Expr::ClassRef(name)` — class identifier referenced as a value
-///     (the lowering at `crates/perry-hir/src/lower.rs::ast::Expr::Ident`
-///     turns class names referenced as values into ClassRef so they
-///     can flow through generic Expr slots without losing the class
-///     identity).
-///   - `Expr::PropertyGet { object: GlobalGet(_), property }` — a
-///     property access on the global object, e.g. `globalThis.WebSocket`
-///     or `window.Date`. The `globalThis.X` form is what the parser
-///     emits for `new globalThis.WebSocket(url)` (mango uses this for
-///     the websocket helper in `_wsOpen`).
-///   - `Expr::PropertyGet { object: LocalGet(ns_id), property }` where
-///     `ns_id` is a namespace import local (`import * as ns from 'm';
-///     new ns.Foo()`). The local id is mapped to its name via
-///     `ctx.local_id_to_name`, then checked against
-///     `ctx.namespace_imports`. The property name is returned as the
-///     class name; the rest of the lower_new path resolves it via the
-///     usual `ctx.classes` lookup, which contains imported classes
-///     under their original (un-namespaced) names.
-fn try_static_class_name<'a>(callee: &'a Expr, ctx: &FnCtx<'_>) -> Option<&'a str> {
-    match callee {
-        Expr::ClassRef(name) => Some(name.as_str()),
-        // Refs #486: `new _X()` where `_X` is the inner self-binding name of
-        // a class expression (e.g. `var X = class _X { ... new _X() ... }`)
-        // lowers to `NewDynamic { callee: ExternFuncRef("_X") }` because the
-        // inner name isn't a real outer-scope identifier — the HIR walker
-        // can't resolve it to anything but an unknown extern. Recognize it
-        // here by checking the per-module class_ids table, which codegen has
-        // already populated with the inner-name → same-id mapping at
-        // compile_module entry. Without this, the call falls through to the
-        // empty-object placeholder path with class_id=0 and method dispatch
-        // breaks on the resulting instance.
-        Expr::ExternFuncRef { name, .. } if ctx.class_ids.contains_key(name) => Some(name.as_str()),
-        Expr::PropertyGet { object, property } => {
-            if matches!(object.as_ref(), Expr::GlobalGet(_)) {
-                return Some(property.as_str());
-            }
-            // Namespace import via local: `import * as ns from 'm'; new ns.Foo()`.
-            // The local binding shows up as `LocalGet(id)` here; we map id →
-            // name via `local_id_to_name`, then check `namespace_imports`.
-            if let Expr::LocalGet(id) = object.as_ref() {
-                if let Some(name) = ctx.local_id_to_name.get(id) {
-                    if ctx.namespace_imports.contains(name) {
-                        return Some(property.as_str());
-                    }
-                }
-            }
-            // Namespace import via ExternFuncRef: the HIR's
-            // `ast::Expr::Ident` lowering at `crates/perry-hir/src/lower.rs`
-            // lifts a namespace identifier to `Expr::ExternFuncRef { name: "ns" }`
-            // when the name resolves to a `import * as ns from 'm'` binding
-            // (rather than a local let). The property access then becomes
-            // `PropertyGet { object: ExternFuncRef("ns"), property: "Foo" }`.
-            // Check `namespace_imports` directly with the ExternFuncRef name.
-            if let Expr::ExternFuncRef { name, .. } = object.as_ref() {
-                if ctx.namespace_imports.contains(name) {
-                    return Some(property.as_str());
-                }
-            }
-            None
-        }
-        _ => None,
-    }
-}
-
-/// Alias kept for backwards compatibility with existing callers
-/// in `stmt.rs` and `codegen.rs` that use the `_pub` suffix.
-pub(crate) fn nanbox_pointer_inline_pub(blk: &mut LlBlock, ptr_i64: &str) -> String {
-    nanbox_pointer_inline(blk, ptr_i64)
-}
-
-/// Inline NaN-box of a raw string handle with `STRING_TAG`.
-pub(crate) fn nanbox_string_inline(blk: &mut LlBlock, ptr_i64: &str) -> String {
-    let tagged = blk.or(I64, ptr_i64, STRING_TAG_I64);
-    blk.bitcast_i64_to_double(&tagged)
-}
-
-/// Convert an i32 boolean (0 or 1) returned by a runtime function into a
-/// NaN-tagged JSValue boolean (`TAG_TRUE` / `TAG_FALSE`).
-pub(crate) fn i32_bool_to_nanbox(blk: &mut LlBlock, i32_val: &str) -> String {
-    let bit = blk.icmp_ne(I32, i32_val, "0");
-    let tagged = blk.select(
-        I1,
-        &bit,
-        I64,
-        crate::nanbox::TAG_TRUE_I64,
-        crate::nanbox::TAG_FALSE_I64,
-    );
-    blk.bitcast_i64_to_double(&tagged)
-}
 /// Per-function codegen context. Held briefly during lowering, never stored.
 pub(crate) struct FnCtx<'a> {
     /// Function being built (blocks, params, registers).
